feat(ci, cli, contracts): update CI workflows, command registry, and enhance formatting commands
This commit enhances the CI configuration by updating the Clippy command to include all targets for comprehensive linting. The command registry has been modified to retire the `vox install` command, replacing it with `vox add`, `vox lock`, and `vox sync` for package management. Additionally, the `vox fmt` command has been wired to format `.vox` files in place, with an option to check formatting without writing changes. The `.gitignore` file has been updated to exclude new checkpoint state files, and various documentation updates have been made to reflect these changes, ensuring a cohesive user experience.

+++ crates/vox-ludus/src/reward_policy.rs
@@ -253,7 +253,7 @@ pub fn learning_mode_crystal_jitter(
     user_id.hash(&mut h);
     day.hash(&mut h);
     event_type.hash(&mut h);
-    (h.finish() % 4) as u64
+    h.finish() % 4
 }
 
 // ── Policy engine ─────────────────────────────────────────
